==> scripts/03_All Taxa Plots.R <==
####################################################################
############# PLOT OF ALL TAXA TOTALS ALL SAMPLES ##################
####################################################################
### 11/23/22

### Example ideas from meeting with Wim 11/22/22

#x <-readf()
#head(x)
#x <- select(x, group=grp_typ, totalCPM = total_cts_per_ml)
#x$seq <- factor(1:nrow(x), ordered=T, labels=abbreviate(x$group))


#p <- ggplot(x, aes(x=seq, totalCPM)) +
  #geom_point(size=1, color="blue") +
 # scale_x_discrete ("") +
# scale_y_log10("Total Counts per Ml, all samples") +
 # wimGraph()
#p

# p +  theme(axis.text.x = element_text(angle=90))
